Switch to using "mvn deploy"
For MTA projects, mvn deploy:deploy-file is used
to deploy yaml and mtar files directly. This still
results in maven-metadata.xml to be maintained on
the nexus. For Maven projects, no assumptions about
project structure are used anymore at all, it is all
up to maven. There are several TODOs left:
 - Providing local .m2 directory to the maven options
   when deploying a maven project
 - Generating a repository entry with credentials
   in the local maven settings, in such a way that
   the properties are substituted from ENV variables.
 - Fix all the Unit Tests.